embed: isolate ONNX/CGO behind localassets build tag
Stage 1 of the embedding/adoption roadmap. CGO-free builds
(CGO_ENABLED=0, gomobile, WASM) now compile cleanly without
requiring ONNX Runtime.

Changes:
- Add //go:build localassets to local.go — isolates the
  yalue/onnxruntime_go CGO dependency behind the existing
  localassets tag
- Extract clsPool/meanPool/l2Normalize to local_math.go
  (no build tag, no CGO) — pure math helpers always available
- Add LocalProvider stub to local_assets_noembed.go — satisfies
  the Provider interface in non-localassets builds; Init returns
  a clear error directing callers to use a network provider
- Split local_test.go: math tests move to local_math_test.go
  (always runs), LocalProvider/utility tests keep localassets tag

CGO_ENABLED=0 go build ./... now succeeds.
Daemon builds with -tags localassets are byte-for-byte identical.
